backend/requirements.txt
Django     ==        4.2.5
django-cors-headers == 4.2.0
djangorestframework == 3.14.0
python-dotenv   ==    1.0.0
Pillow       ==       10.0.1
redis == 5.0.0
django-redis==4.12.1
factory_boy == 3.3.0